Invoice functionality added

+++ controllers/admin.js
@@ -1,5 +1,6 @@
 const Product = require('../models/product');
-const product = require('../models/product');
+
+const fileHelper = require('../util/file');
 
 const { validationResult } = require('express-validator/check')
 
@@ -17,9 +18,24 @@ exports.getAddProduct = (req, res, next) => {
 
 exports.postAddProduct = (req, res, next) => {
   const title = req.body.title;
-  const imageUrl = req.body.imageUrl;
+  const image = req.file;
   const price = req.body.price;
   const description = req.body.description;
+  if (!image){
+    return res.status(422).render('admin/edit-product', {
+      pageTitle: 'Add Product',
+      path: '/admin/add-product',
+      editing: false,
+      hasError: true,
+      product: {
+        title: title,
+        price: price,
+        description: description
+      },
+      errorMessage: 'Attached File is not an image.',
+      validationErrors: []
+    });
+  }
   const errors = validationResult(req);
 
   if(!errors.isEmpty()){
@@ -30,7 +46,6 @@ exports.postAddProduct = (req, res, next) => {
       hasError: true,
       product: {
         title: title,
-        imageUrl: imageUrl,
         price: price,
         description: description
       },
@@ -38,6 +53,8 @@ exports.postAddProduct = (req, res, next) => {
       validationErrors: errors.array()
     });
   }
+
+  const imageUrl = image.path;
 
   const product = new Product({
     title: title, 
@@ -69,7 +86,7 @@ exports.postAddProduct = (req, res, next) => {
       //   errorMessage: "Database operation failed, please try again.",
       //   validationErrors: []
       // });
-      const error = new Error(error);
+      const error = new Error(err);
       error.httpStatusCode = 500;
       return next(error);
     });
@@ -82,7 +99,6 @@ exports.getEditProduct = (req, res, next) => {
   }
   const prodId = req.params.productId;
   Product.findById(prodId)
-    // Product.findById(prodId)
     .then(product => {
       // throw new Error('dummy!');
       if (!product) {
@@ -99,6 +115,7 @@ exports.getEditProduct = (req, res, next) => {
       });
     })
     .catch(err => {
+      console.log(err);
       const error = new Error(err) ;
       error.httpStatusCode = 500;
       next(error);
@@ -109,7 +126,7 @@ exports.postEditProduct = (req, res, next) => {
   const prodId = req.body.productId;
   const updatedTitle = req.body.title;
   const updatedPrice = req.body.price;
-  const updatedImageUrl = req.body.imageUrl;
+  const image = req.file;
   const updatedDesc = req.body.description;
 
   const errors = validationResult(req);
@@ -122,7 +139,6 @@ exports.postEditProduct = (req, res, next) => {
       hasError: true,
       product: {
         title: updatedTitle,
-        imageUrl: updatedImageUrl,
         price: updatedPrice,
         description: updatedDesc,
         _id: prodId
@@ -141,7 +157,10 @@ exports.postEditProduct = (req, res, next) => {
     product.title = updatedTitle;
     product.price = updatedPrice;
     product.description = updatedDesc;
-    product.imageUrl = updatedImageUrl;
+    if (image) {
+      fileHelper.deleteFile(product.imageUrl);
+      product.imageUrl = image.path;
+    }
     return product.save()
     .then(result => {
       console.log('UPDATED PRODUCT!');
@@ -176,7 +195,14 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
   const prodId = req.body.productId;
-  Product.deleteOne({ _id: prodId, userId: req.user._id})
+  Product.findById(prodId)
+    .then((product) => {
+      if(!product) {
+        return next(new Error('Product not found.'))
+      }
+      fileHelper.deleteFile(product.imageUrl);
+      return Product.deleteOne({ _id: prodId, userId: req.user._id})
+    })
     .then(result => {
       console.log('DESTROYED PRODUCT');
       res.redirect('/admin/products');
